+++ Assets/theArchitectTechPack.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 317e24cfb64af4f43a7e31d4d04dfea1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
