+++ Assets/theArchitectTechPack/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa493391dfb2467428cf5ba44f95533e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
