+++ Assets/theArchitectTechPack/theArchTechMainAsset.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "theArchTechMainAsset", menuName = "theArchTechPack/theArchTechMainAsset", order = 51)]
+public class theArchTechMainAsset : ScriptableObject
+{
+    
+}
