+++ Assets/theArchitectTechPack/theArchTechMainAsset.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93380cd7a7fe4364f98a9c355f59edca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
